requirements.txt
colorama==0.3.7
decorator==4.0.10
Django==1.10.2
ipython==5.1.0
ipython-genutils==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.7
Pygments==2.1.3
pyserial==3.1.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
win-unicode-console==0.5
